requirements/requirements.txt
yfinance
pandas
numpy
psycopg2-binary
scipy
scikit-learn
matplotlib
seaborn